requirements.txt
absl-py==0.12.0
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.4
astunparse==1.6.3
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.2
celery==4.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cmake==3.18.4.post1
cycler==0.10.0
Django==3.2.3
django-background-tasks==1.2.5
django-celery-beat==2.2.0
django-cleanup==5.2.0
django-compat==1.0.15
django-timezone-field==4.1.2
dlib==19.22.0
dnspython==1.16.0
eventlet==0.31.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==4.0.1
imutils==0.5.4
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.11
Markdown==3.3.4
matplotlib==3.3.4
numpy @ file:///D:/bld/numpy_1610324703282/work
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
